restrict target words to 8 bytes

+++ test/mwcas_entry_test.cpp
@@ -27,15 +27,15 @@ class MwCASEntryFixture : public ::testing::Test
 
 TEST_F(MwCASEntryFixture, Construct_InitialValues_MemberVariablesCorrectlyInitialized)
 {
-  auto target = uint32_t{10};
+  auto target = uint64_t{10};
   const auto old_val = target;
-  const auto new_val = uint32_t{20};
+  const auto new_val = uint64_t{20};
 
   const auto entry = MwCASEntry{&target, old_val, new_val};
 
   EXPECT_EQ(static_cast<void*>(&target), static_cast<void*>(entry.addr));
-  EXPECT_EQ(old_val, entry.old_val.GetTargetData<uint32_t>());
-  EXPECT_EQ(new_val, entry.new_val.GetTargetData<uint32_t>());
+  EXPECT_EQ(old_val, entry.old_val.GetTargetData<uint64_t>());
+  EXPECT_EQ(new_val, entry.new_val.GetTargetData<uint64_t>());
 }
 
 }  // namespace dbgroup::atomic::mwcas
